--- Packages/io.chainsafe.web3-unity/Runtime/Scripts/Samples/GelatoSample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3f0896c2e77ea734398b3420168db1e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
