+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.6.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/Erc20/Erc20Calls.cs
@@ -47,7 +47,7 @@ public class Erc20Calls : SampleBase<Erc20Calls>, IWeb3InitializedHandler, ILogo
     #endregion
 
     private Erc20Contract _erc20;
-
+    
     /// <summary>
     /// Balance Of ERC20 Address
     /// </summary>
