+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.6.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/Gelato/GelatoCalls.cs
@@ -1,63 +1,224 @@
+using System.Numerics;
 using System.Threading.Tasks;
-using ChainSafe.Gaming.UnityPackage.Connection;
 using ChainSafe.Gaming.Web3;
+using ChainSafe.GamingSdk.Gelato;
+using ChainSafe.GamingSdk.Gelato.Dto;
+using ChainSafe.Gaming.UnityPackage.Connection;
 using ChainSafe.Gaming.Web3.Build;
 using Microsoft.Extensions.DependencyInjection;
 
 public class GelatoCalls : SampleBase<GelatoCalls>, IWeb3InitializedHandler
 {
-    // Fields
-    private GelatoSample gelato;
+    private Web3 _web3;
 
     /// <summary>
     /// Gelato with sync fee
     /// </summary>
-    public async Task<string> GelatoCallWithSyncFee()
+    public async Task<string> CallWithSyncFee()
     {
-        var result = await gelato.CallWithSyncFee();
-        return SampleOutputUtil.BuildResultMessage(
-            $"Task complete. Final status of {result.TaskId}: {result.Status.TaskState}. " +
-            $"Transaction hash: {result.Status.TransactionHash}",
-            nameof(GelatoSample), nameof(GelatoSample.CallWithSyncFee));
-    }
+        const string vitalik = "0xd8da6bf26964af9d7eed9e03e53415d37aa96045";
+        const string target = "0xA045eb75e78f4988d42c3cd201365bDD5D76D406";
+        const string feeToken = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
+        const string abi = "[{\"inputs\": [" +
+                           "{\"internalType\":\"address\",\"name\":\"_token\",\"type\":\"address\"}," +
+                           "{\"internalType\":\"address\",\"name\":\"_to\",\"type\":\"address\"}," +
+                           "{\"internalType\":\"uint256\",\"name\":\"_amount\",\"type\":\"uint256\"}" +
+                           "]," +
+                           "\"name\":\"sendToFriend\"," +
+                           "\"outputs\":[]," +
+                           "\"stateMutability\":\"nonpayable\"," +
+                           "\"type\":\"function\"" +
+                           "}]";
+        var contract = _web3.ContractBuilder.Build(abi, target);
+        var data = contract.Calldata("sendToFriend", new object[]
+        {
+            feeToken,
+            vitalik,
+            new BigInteger(5 * 10E12),
+        });
 
-    /// <summary>
-    /// Gelato2771 with sync fee
-    /// </summary>
-    public async Task<string> GelatoCallWithSyncFeeErc2771()
-    {
-        var result = await gelato.CallWithSyncFeeErc2771();
-        return SampleOutputUtil.BuildResultMessage(
-            $"Task complete. Final status of {result.TaskId}: {result.Status.TaskState}. " +
-            $"Transaction hash: {result.Status.TransactionHash}",
-            nameof(GelatoSample), nameof(GelatoSample.CallWithSyncFeeErc2771));
+        var relayResponse = await _web3.Gelato().CallWithSyncFee(new CallWithSyncFeeRequest()
+        {
+            Data = data,
+            FeeToken = feeToken,
+            IsRelayContext = true,
+            Target = target,
+        });
+
+        while (true)
+        {
+            var status = await _web3.Gelato().GetTaskStatus(relayResponse.TaskId);
+
+            switch (status.TaskState)
+            {
+                case TaskState.ExecSuccess:
+                case TaskState.ExecReverted:
+                case TaskState.Cancelled:
+                    return SampleOutputUtil.BuildResultMessage(
+                        $"Task complete. Final status of {relayResponse.TaskId}: {status.TaskState}. " +
+                        $"Transaction hash: {status.TransactionHash}",
+                        nameof(GelatoCalls), nameof(CallWithSyncFee));
+                default:
+                    await WaitForSeconds(2);
+                    continue;
+            }
+        }
     }
 
     /// <summary>
     /// Gelato sponsor call
     /// </summary>
-    public async Task<string> GelatoSponsorCall()
+    public async Task<string> SponsorCall()
     {
-        var result = await gelato.SponsorCall();
-        return SampleOutputUtil.BuildResultMessage(
-            $"Task complete. Final status of {result.TaskId}: {result.Status.TaskState}. " +
-            $"Transaction hash: {result.Status.TransactionHash}",
-            nameof(GelatoSample), nameof(GelatoSample.SponsorCall));
+        const string counterContract = "0x763D37aB388C5cdd2Fb0849d6275802F959fbF30";
+
+        const string abi = "[{\"inputs\": []," +
+                           "\"name\":\"increment\"," +
+                           "\"outputs\":[]," +
+                           "\"stateMutability\":\"nonpayable\"," +
+                           "\"type\":\"function\"" +
+                           "}]";
+        var contract = _web3.ContractBuilder.Build(abi, counterContract);
+
+        var data = contract.Calldata("increment");
+
+        var relayResponse = await _web3.Gelato().SponsoredCall(new SponsoredCallRequest()
+        {
+            Target = counterContract,
+            Data = data,
+        });
+
+        while (true)
+        {
+            var status = await _web3.Gelato().GetTaskStatus(relayResponse.TaskId);
+
+            switch (status.TaskState)
+            {
+                case TaskState.ExecSuccess:
+                case TaskState.ExecReverted:
+                case TaskState.Cancelled:
+                    return SampleOutputUtil.BuildResultMessage(
+                        $"Task complete. Final status of {relayResponse.TaskId}: {status.TaskState}. " +
+                        $"Transaction hash: {status.TransactionHash}",
+                        nameof(GelatoCalls), nameof(SponsorCall));
+                default:
+                    await WaitForSeconds(2);
+                    continue;
+            }
+        }
+    }
+
+    /// <summary>
+    /// Gelato2771 with sync fee
+    /// </summary>
+    public async Task<string> CallWithSyncFeeErc2771()
+    {
+        const string target = "0x5dD1100f23278e0e27972eacb4F1B81D97D071B7";
+        const string feeToken = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
+        const string abi = "[{\"inputs\": []," +
+                           "\"name\":\"increment\"," +
+                           "\"outputs\":[]," +
+                           "\"stateMutability\":\"nonpayable\"," +
+                           "\"type\":\"function\"" +
+                           "}]";
+        var contract = _web3.ContractBuilder.Build(abi, target);
+
+        var data = contract.Calldata("increment", new object[]
+        {
+        });
+
+        var relayResponse = await _web3.Gelato().CallWithSyncFeeErc2771(new CallWithSyncFeeErc2771Request()
+        {
+            Target = target,
+            Data = data,
+            FeeToken = feeToken,
+            IsRelayContext = true,
+        });
+
+        while (true)
+        {
+            var status = await _web3.Gelato().GetTaskStatus(relayResponse.TaskId);
+
+            switch (status.TaskState)
+            {
+                case TaskState.ExecSuccess:
+                case TaskState.ExecReverted:
+                case TaskState.Cancelled:
+                    return SampleOutputUtil.BuildResultMessage(
+                        $"Task complete. Final status of {status.TaskId}: {status.TaskState}. " +
+                        $"Transaction hash: {status.TransactionHash}",
+                        nameof(GelatoCalls), nameof(CallWithSyncFeeErc2771));
+                default:
+                    await WaitForSeconds(2);
+                    continue;
+            }
+        }
     }
 
     /// <summary>
     /// Gelato2771 sponsor call
     /// </summary>
-    public async Task<string> GelatoSponsorCallErc2771()
+    public async Task<string> SponsorCallErc2771()
     {
-        var result = await gelato.SponsorCallErc2771();
+        const string target = "0x00172f67db60E5fA346e599cdE675f0ca213b47b";
 
-        return SampleOutputUtil.BuildResultMessage(
-            $"Task complete. Final status of {result.TaskId}: {result.Status.TaskState}. " +
-            $"Transaction hash: {result.Status.TransactionHash}",
-            nameof(GelatoSample), nameof(GelatoSample.SponsorCallErc2771));
+        const string abi = "[{\"inputs\": []," +
+                  "\"name\":\"increment\"," +
+                  "\"outputs\":[]," +
+                  "\"stateMutability\":\"nonpayable\"," +
+                  "\"type\":\"function\"" +
+                  "}]";
+
+        var contract = _web3.ContractBuilder.Build(abi, target);
+
+        var data = contract.Calldata("increment");
+
+        var relayResponse = await _web3.Gelato().SponsoredCallErc2771(new SponsoredCallErc2771Request()
+        {
+            Target = target,
+            Data = data,
+            User = _web3.Signer.PublicAddress,
+        });
+
+        while (true)
+        {
+            var status = await _web3.Gelato().GetTaskStatus(relayResponse.TaskId);
+
+            switch (status.TaskState)
+            {
+                case TaskState.ExecSuccess:
+                case TaskState.ExecReverted:
+                case TaskState.Cancelled:
+                    return SampleOutputUtil.BuildResultMessage(
+                        $"Task complete. Final status of {status.TaskId}: {status.TaskState}. " +
+                        $"Transaction hash: {status.TransactionHash}",
+                        nameof(GelatoCalls), nameof(SponsorCallErc2771));
+                default:
+                    await WaitForSeconds(2);
+                    continue;
+            }
+        }
     }
 
+    public bool GetGelatoDisabled()
+    {
+        return _web3.Gelato().GetGelatoDisabled();
+    }
+
+    async Task WaitForSeconds(int seconds)
+    {
+        // Task.Delay doesn't work on WebGL
+#if UNITY_WEBGL && !UNITY_EDITOR
+        var now = Time.time;
+        while (Time.time - now < 2)
+        {
+            await Task.Yield();
+        }
+#else
+        await Task.Delay(seconds * 1000);
+#endif
+    }
+    
     public override Web3Builder ConfigureServices(Web3Builder web3Builder)
     {
         return web3Builder.Configure(services =>
@@ -68,8 +229,8 @@ public class GelatoCalls : SampleBase<GelatoCalls>, IWeb3InitializedHandler
 
     public Task OnWeb3Initialized(Web3 web3)
     {
-        gelato = new GelatoSample(web3);
-
+        _web3 = web3;
+        
         return Task.CompletedTask;
     }
 }
