--- src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.6.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/Gelato/GelatoDisableContent.cs
@@ -1,15 +1,0 @@
-using ChainSafe.Gaming.UnityPackage;
-using UnityEngine;
-
-public class GelatoDisableContent : MonoBehaviour
-{
-    private GelatoSample gelato;
-
-    private void Awake()
-    {
-        gelato = new GelatoSample(Web3Unity.Web3);
-        if (!gelato.GetGelatoDisabled()) return;
-        Debug.Log("Gelato functionality disabled as your chain isn't supported");
-        gameObject.SetActive(false);
-    }
-}
