--- src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.6.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/Gelato/GelatoDisableContent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f404484bf3c7d4246a08e60ab1cd69b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
