+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.6.0/Web3.Unity Samples/Scripts/Scenes/SampleMain/IPFS/IPFSCalls.cs
@@ -83,9 +83,4 @@ public class IPFSCalls : SampleBase<IPFSCalls>
     }
 
     #endregion
-
-    public override Web3Builder ConfigureServices(Web3Builder web3Builder)
-    {
-        return web3Builder;
-    }
 }
